Refactoring
- Bot manager moved to pipobot/__init__.py. A __main__.py was added to the bot
can be started with python -m pipobot [options] (for Python >= 2.7; on older
versions, python -m pipobot.__main__ [options] can be used)
- Command line and configuration parsing was moved to a Configuration class in
the pipobot.config module. The configuration format was slightly changed to be
more compact. A sample configuration file was added so it can be automatically
installed into the system.
- gettext-related functions were moved to the pipobot.translations module.
- User-management functions were moved from pipobot.lib.user to pipobot.user.
- The bot core was moved from pipobot.bot_jabber to pipobot.xmpp_bot.
- Bot module bases were moved from pipobot.lib.modules to the pipobot.modules
package.
- Updated MANIFEST.in